tweak
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@12450 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_xml_config.c
@@ -181,7 +181,7 @@ SWITCH_DECLARE(switch_status_t) switch_xml_config_parse(switch_xml_t xml, int re
 			case SWITCH_CONFIG_FLAGARRAY:
 				{
 					int8_t *dest = (int8_t*)item->ptr;
-					int8_t index = (int)(intptr_t)item->data;
+					int8_t index = (int8_t)(intptr_t)item->data;
 					int8_t newval = value ? !!switch_true(value) : (int8_t)((intptr_t)item->defaultvalue);
 					if (dest[index] != newval) {
 						changed = SWITCH_TRUE;
